+++ Cargo.toml
@@ -5,6 +5,6 @@ edition = "2021"
 
 [dependencies]
 rand = "0.8.5"
-clap = "4.5.26"
+clap = "4.5.27"
 rodio = "0.20.1"
 termion = "4.0.3"
