+++ src/application/usecases/mod.rs
@@ -1,0 +1,2 @@
+pub mod generate;
+pub mod user_input;
